+++ src/Video/VideoInfoReaderFactory.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Soluble\MediaTools\Video;
 
 use Psr\Container\ContainerInterface;
+use Soluble\MediaTools\Video\Cache\CacheInterface;
 use Soluble\MediaTools\Video\Config\FFProbeConfigInterface;
 use Soluble\MediaTools\Video\Logger\LoggerInterface;
 
@@ -20,10 +21,12 @@ class VideoInfoReaderFactory
     public function __invoke(ContainerInterface $container): VideoInfoReaderInterface
     {
         $logger = $container->has(LoggerInterface::class) ? $container->get(LoggerInterface::class) : null;
+        $cache  = $container->has(CacheInterface::class) ? $container->get(CacheInterface::class) : null;
 
         return new VideoInfoReader(
             $container->get(FFProbeConfigInterface::class),
-            $logger
+            $logger,
+            $cache
         );
     }
 }